Perform analysis after filtering out data that doesn't need to be inspected
